fix: variant type

+++ packages/node-opcua-address-space/source/set_namespace_meta_data.ts
@@ -54,7 +54,7 @@ export function setNamespaceMetaData(namespace: INamespace): void {
         {
             get: () =>
                 new Variant({
-                    dataType: DataType.UInt32,
+                    dataType: DataType.UInt16,
                     value: namespace.getDefaultAccessRestrictions()
                 })
         },
